refactor(kamn-node): extract report rendering module from main (#2488)
